fix : deny access unless connected as Beekeeper

+++ src/Controller/ApiaryController.php
@@ -6,6 +6,7 @@ use App\Entity\Apiary;
 use App\Form\ApiaryType;
 use App\Repository\ApiaryRepository;
 use App\Repository\BeehiveRepository;
+use App\Repository\BeekeeperRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -16,8 +17,9 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class ApiaryController extends AbstractController
 {
     #[Route('/', name: 'app_apiary_index', methods: ['GET'])]
-    public function index(ApiaryRepository $apiaryRepository): Response
+    public function index(ApiaryRepository $apiaryRepository, BeekeeperRepository $beekeeperRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_BEEKEEPER');
         $user = $this->getUser();
         $apiaries = $apiaryRepository->findApiariesByBeekeeper($user->getId());
 
@@ -27,8 +29,9 @@ class ApiaryController extends AbstractController
     }
 
     #[Route('/new', name: 'app_apiary_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $em): Response
+    public function new(Request $request, EntityManagerInterface $em, BeekeeperRepository $beekeeperRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_BEEKEEPER');
         $apiary = new Apiary();
         $form = $this->createForm(ApiaryType::class, $apiary);
         $form->handleRequest($request);
@@ -51,6 +54,8 @@ class ApiaryController extends AbstractController
     #[Route('/{id}', name: 'app_beehive_by_apiary_show', methods: ['GET'])]
     public function show(Apiary $apiary, BeehiveRepository $beehiveRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_BEEKEEPER');
+
         $beehives = $beehiveRepository->findBy([
             'apiary' => $apiary->getId(),
         ]);
@@ -64,6 +69,8 @@ class ApiaryController extends AbstractController
     #[Route('/{id}/edit', name: 'app_apiary_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Apiary $apiary, ApiaryRepository $apiaryRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_BEEKEEPER');
+
         $form = $this->createForm(ApiaryType::class, $apiary);
         $form->handleRequest($request);
 
@@ -82,6 +89,8 @@ class ApiaryController extends AbstractController
     #[Route('/{id}/delete', name: 'app_apiary_delete', methods: ['GET'])]
     public function delete(Apiary $apiary, ApiaryRepository $apiaryRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_BEEKEEPER');
+
         $apiaryRepository->remove($apiary, true);
 
         return $this->redirectToRoute('app_apiary_index', [], Response::HTTP_SEE_OTHER);
